requirements.txt
sounddevice
soundfile
soundcard
numpy
faster-whisper
pyannote.audio
